requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# ML/NLP
torch>=2.0.0
transformers>=4.35.0

# Document Processing
pypdf>=3.17.0

# Data & Visualization
pandas>=2.0.0
plotly>=5.18.0
